API Update
- Updated order route to check for token on session rather than just
session itself.
- Same as above but for users route.
- Added /reauthenticate to the users route; allows a user to
reauthenticate their login using a token.
- Login now only stores token, token expiry and security in the session.
- Added token and token expiry keys to the user model. These aren't
required. Token isn't unique because having it not required would break
this, and because the way we generate a token is unique.
- Removed the 'save' pre from the model as it was breaking when saving
the token; it would re-hash the hashed password.
- Updated the authenticate static so that it now generates and returns
the login token and token expiry. At the same time it also saves these
pieces of information agains the user in the database.
- Added a reauthenticate static that takes a token and checks to see if
there is a valid token in the users database, if there is then it does
an expiry check and only returns the user if both are valid.
- Moved password encryption on user creation to the user controller from
the save pre.
- Updated login in the controller to now return the token and token
expiry.
- Added a reauthenticate method to the user controller. Takes a token
and basically just calls the reauthenticate static.