App Component Update
- Changed the HTML file to be a <router-outlet> for our routing.
- AppComponent constructor is used to decide where to direct the user
when they first load the website.
- It subscribes to the user service and when a user is set this is where
it will direct them to their home area. It will also redirect to login
when they are unset (but this was removed for now cause bugs).
- When we first load the page (or if we've refreshed) it will also check
the session storage to see if a token was set, if the token was set it
does some date checking and then checks to see if the token is valid.
This needs to be changed though as the end point also does date
checking, no point doing it twice!