Order API Update
- Modified the 'add' function in the Order controller to turn it into an
async function. It takes stock first, if that failed it rolls back, if
not then it places the order. If the order failed it rolls back,
otherwise it returns the order.
- Updated Order model to make the minimum value of friendly ID 1. Also
added a 'place' static function that handles placing the order.
- Updated OrderStatus model to make the value unique.
- Updated Stock model to make the quantity a number and to give it a
minimum value of 1.
- Removed 'give' and 'take' functions from the Stock model and replaced
it with an 'adjust' function. This will adjust the stock positively or
negatively.
- Updated Order route to make use of the new async function.