Updated Users API
**Improvements**
- Added documentation to all functions (including param flags).
- Added hook into the model to capture when the password value is
changed against a user, and if the password differs from what is
currently set then they are setting a new password so we encrypt it.
This way we don't have to encrypt elsewhere.
- Re-wrote the three static methods the be more robust.
- Re-wrote the controller to also be more robust.
- Renamed controller methods and added place holders for 'many' update
and delete functions.
- Moved the obtaining of orders for a user to this controller and will
remove from the orders controller.
- Added filtering by username, homepath and admin flag to the 'many'
function when searching for users.
- Added the /user/:id/orders route to obtain orders for a specific user.
- Tweaked the route file to make use of the new security class and to
send proper HTTP status codes.

**Fixes**
- User homepath and admin flag cannot be updated unless the person doing
the update is an admin. Password can still be changed by anyone as long
as they are the user or an admin.
- There were some issues with the way users were being updated and
logged in. This is now fixed thanks to the re-writes.