Updated Orders API
**Improvements**
- Added documentation to all functions (including param flags).
- Added mongoose-double package to handle Double numbers properly.
- Removed the route that obtains the list of orders for a user as that
is now part of the users section of the API.
- Updated all routes to make use of the new way security works.
- Added status key to the content sub-item in the model.
- Added value key to the model.
- Changed a few times, added some required keys.
- Date gets generated using the default value. It gets set to the
current date. Previously we were passing it in from the front-end.
- Added a pre save hook that will generate the price When the document
is about to be saved it will iterate through the list of items that we
have in the order, and then it get the price for each and tally it.
- Re-wrote the place and nextId functions to make them more robust.
- Updated the update function in the controller to make sure we can only
explicitly update content, status and table ID. We don't want any other
values to change (value will still be auto-generated).
- Re-wrote the other functions to make them more robust and to follow
new standards.

**Fixes**
- Error message in users API changed.